requirements.txt
fastapi
sqlalchemy
uvicorn
psycopg2
python-jose[cryptography]
passlib[bcrypt]
python-multipart